requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
requests>=2.31.0
httpx>=0.24.0
socksio>=1.0.0
pysocks>=1.7.1
